Reorganise world unit tests
Remove unused import
